requirements.txt
requests
simplejson
bcrypt
flask
twilio
python-dotenv
bcrypt
mongoengine
bs4
sendgrid
stripe
python-dateutil
boto
flask-cors
flask-jwt-extended
flask-restx
pytest
python-dateutil
authy
faker
flask-socketio
python-socketio[client]
gevent
eventlet
dnspython
validate_email
pika
